start_with/start_with-main/server/scripts/requirements.txt
# Requirements for PDF text extraction
# Install with: pip install -r requirements.txt

# PyMuPDF - Fast PDF processing (text extraction + rendering)
PyMuPDF>=1.24.0

# Pillow - Image processing (required for pytesseract)
Pillow>=10.0.0

# pytesseract - Python wrapper for Tesseract OCR
pytesseract>=0.3.10

# Note: You also need Tesseract OCR installed on the system:
# Ubuntu/Debian: sudo apt-get install tesseract-ocr tesseract-ocr-eng
# macOS: brew install tesseract
# Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
